# specs/001-specify-scripts-bash/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: FT Backend Core Services API
  version: 0.1.0
servers:
  - url: https://localhost:3000
    description: Development via Docker Compose
paths:
...
  /users/{id}:
    get:
      summary: Fetch public profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
  /users/{id}/stats:
    get:
      summary: Fetch match statistics
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Stats payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStats'
...
  /matches:
    get:
      summary: List matches (filters by player or status)
      security:
        - bearerAuth: []
      parameters:
        - name: playerId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchRecord'
    post:
      summary: Report match result
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchReport'
      responses:
        '200':
          description: Match updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchRecord'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
        bio:
          type: string
      required: [id, displayName]
    UserStats:
      type: object
      properties:
        wins:
          type: integer
        losses:
          type: integer
        currentStreak:
          type: integer
        longestStreak:
          type: integer
        lastMatchAt:
          type: string
          format: date-time
...
    MatchRecord:
      type: object
      properties:
        id:
          type: string
        playerOneId:
          type: string
        playerTwoId:
          type: string
        status:
          type: string
        winnerId:
          type: string
          nullable: true
      required: [id, playerOneId, playerTwoId, status]
    MatchReport:
      type: object
      properties:
        matchId:
          type: string
        winnerId:
          type: string
      required: [matchId, winnerId]
...